Issue/141 memmap

+++ linux-4.9.3/kernel/s2e/s2e.c
@@ -30,9 +30,9 @@
 
 static int __init s2e_init(void)
 {
-	size_t current_task_addr = &s2e_current_task;
-	size_t task_struct_pid_offset = offsetof(struct task_struct, pid);
-	size_t task_struct_tgid_offset = offsetof(struct task_struct, tgid);
+	uint64_t current_task_addr = (uintptr_t)&s2e_current_task;
+	uint64_t task_struct_pid_offset = offsetof(struct task_struct, pid);
+	uint64_t task_struct_tgid_offset = offsetof(struct task_struct, tgid);
 
 	if (num_online_cpus() > 1) {
 		s2e_printf("LinuxMonitor only supports single-CPU systems\n");
@@ -41,15 +41,14 @@ static int __init s2e_init(void)
 	}
 
 	/* Check if the LinuxMonitor plugin is enabled */
-	if (boot_cpu_has(X86_FEATURE_S2E)
-	    && s2e_plugin_loaded("LinuxMonitor")) {
+	if (boot_cpu_has(X86_FEATURE_S2E) && s2e_plugin_loaded("LinuxMonitor")) {
 		s2e_linux_monitor_enabled = 1;
 	}
 
 	/* Send addresses and offsets to the LinuxMonitor plugin */
 	if (s2e_linux_monitor_enabled) {
-		s2e_linux_init(PAGE_OFFSET, __START_KERNEL, current_task_addr,
-			       task_struct_pid_offset, task_struct_tgid_offset);
+		s2e_linux_init(PAGE_OFFSET, __START_KERNEL, current_task_addr, task_struct_pid_offset,
+			       task_struct_tgid_offset);
 	}
 
 	return 0;
